+++ kernel/gdt.h
@@ -17,28 +17,28 @@ struct gdt_entry_repr
     // access byte
 
     // must be 0, cpu sets to 1 when it accesses this segment
-    uint8_t  accessed:   1;
+    uint32_t  accessed:  1;
 
     // code: read access
     // data: write access
-    uint8_t  rw:         1;
+    uint32_t  rw:        1;
 
     // code: conforming: if rings <= current can execute this segment
     // data: direction : if this segment grows up
-    uint8_t  dir_conf:   1;
+    uint32_t  dir_conf:  1;
 
     // code: 1
     // data: 0
-    uint8_t  exec:       1;
+    uint32_t  exec:      1;
 
     // who knows
-    uint8_t  one:        1;
+    uint32_t  one:       1;
 
     // ring level 0-3
-    uint8_t  priv:       2;
+    uint32_t  priv:      2;
 
     // must be 1
-    uint8_t  present:    1;
+    uint32_t  present:   1;
 
     // randomly, upper part of the limit
     uint32_t limit_high: 4;
@@ -46,18 +46,18 @@ struct gdt_entry_repr
     // flags
 
     // debugging
-    uint8_t  avl:        1;
+    uint32_t  avl:       1;
 
     // 64 bit mode
-    uint8_t  x86_64:     1;
+    uint32_t  x86_64:    1;
 
     // 32 bit mode: 1
     // 16 bit mode: 0
-    uint8_t  size:       1;
+    uint32_t  size:      1;
 
     // 4KiB pages (allows full 4GiB range): 1
     // 1B blocks                          : 0
-    uint8_t  gran:       1;
+    uint32_t  gran:      1;
 
     // upper part of base
     uint32_t base_high:  8;
